bugfixes and improvements
