+++ utils/consts.js
@@ -1,0 +1,12 @@
+const statusCodes = {
+  error: 500,
+  conflict: 409,
+  forbidden: 403,
+  notFound: 404,
+  badRequest: 400,
+  unauthorized: 401,
+  created: 201,
+  OK: 200,
+};
+
+module.exports = { statusCodes };
